requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
requests==2.27.1
smmap==5.0.0
urllib3==1.26.8
